MS-43: Reorganize useHistory, change Modal to ModalDelete, put all Modal components into a Modal folder, change inputNotes for textAreas in Interviews & Applications, delete extra component in Sessions
