+++ src/Components/Admins/index.js
@@ -7,15 +7,15 @@ import styles from './admins.module.css';
 import ButtonCreate from '../Shared/Buttons/ButtonCreate';
 import ButtonDelete from '../Shared/Buttons/ButtonDelete';
 import ButtonUpdate from '../Shared/Buttons/ButtonUpdate';
-import Modal from '../Shared/Modal';
-import ModalError from '../Shared/ModalError';
+import ModalDelete from '../Shared/Modals/ModalDelete';
+import ModalError from '../Shared/Modals/ModalError';
 
 function Admins() {
+  const history = useHistory();
   const dispatch = useDispatch();
   const admins = useSelector((store) => store.admins.admins);
   const [adminId, setAdminId] = useState([]);
   const [showDelete, setShowDelete] = useState(false);
-  const history = useHistory();
   const isLoading = useSelector((store) => store.admins.isLoading);
   const error = useSelector((store) => store.admins.error);
 
@@ -25,7 +25,7 @@ function Admins() {
 
   return (
     <section className={styles.section}>
-      <Modal
+      <ModalDelete
         show={showDelete}
         title="Delete Admin"
         message="Are you sure you want to delete this Admin?"
