+++ src/Components/Positions/index.js
@@ -7,15 +7,15 @@ import styles from './positions.module.css';
 import ButtonCreate from '../Shared/Buttons/ButtonCreate';
 import ButtonDelete from '../Shared/Buttons/ButtonDelete';
 import ButtonUpdate from '../Shared/Buttons/ButtonUpdate';
-import Modal from '../Shared/Modal';
-import ModalError from '../Shared/ModalError';
+import ModalDelete from '../Shared/Modals/ModalDelete';
+import ModalError from '../Shared/Modals/ModalError';
 
 function Positions() {
+  const history = useHistory();
   const dispatch = useDispatch();
   const positions = useSelector((store) => store.positions.list);
   const [selectedPosition, setSelectedPosition] = useState(undefined);
   const [showDelete, setShowDelete] = useState(false);
-  const history = useHistory();
   const isLoading = useSelector((store) => store.positions.isLoading);
   const error = useSelector((store) => store.positions.error);
 
@@ -33,7 +33,7 @@ function Positions() {
 
   return (
     <section className={styles.section}>
-      <Modal
+      <ModalDelete
         show={showDelete}
         title="Delete a Position"
         message="Are you sure you want to delete this position?"
