+++ src/Components/Profiles/index.js
@@ -7,15 +7,15 @@ import styles from './profiles.module.css';
 import ButtonCreate from '../Shared/Buttons/ButtonCreate';
 import ButtonDelete from '../Shared/Buttons/ButtonDelete';
 import ButtonUpdate from '../Shared/Buttons/ButtonUpdate';
-import Modal from '../Shared/Modal';
-import ModalError from '../Shared/ModalError';
+import ModalDelete from '../Shared/Modals/ModalDelete';
+import ModalError from '../Shared/Modals/ModalError';
 
 function Profiles() {
+  const history = useHistory();
   const dispatch = useDispatch();
   const profiles = useSelector((store) => store.profiles.list);
   const [selectedProfile, setSelectedProfile] = useState(undefined);
   const [showDelete, setShowDelete] = useState(false);
-  const history = useHistory();
   const isLoading = useSelector((store) => store.profiles.isLoading);
   const error = useSelector((store) => store.profiles.error);
 
@@ -33,7 +33,7 @@ function Profiles() {
 
   return (
     <section className={styles.section}>
-      <Modal
+      <ModalDelete
         show={showDelete}
         title="Delete a Profile"
         message="Are you sure you want to delete this profile?"
