+++ src/Components/Psychologists/Form/index.js
@@ -13,9 +13,10 @@ import Input from '../../Shared/Input';
 import Checkbox from '../../Shared/Checkbox';
 import ButtonCancel from '../../Shared/Buttons/ButtonCancel';
 import ButtonConfirm from '../../Shared/Buttons/ButtonConfirm';
-import ModalError from '../../Shared/ModalError';
+import ModalError from '../../Shared/Modals/ModalError';
 
 const Form = () => {
+  const history = useHistory();
   const dispatch = useDispatch();
   const [psychologistId, setPsychologistId] = useState(0);
   const [firstNameForm, setFirstName] = useState('');
@@ -46,7 +47,6 @@ const Form = () => {
   const [fridayTo, setFridayTo] = useState(0);
   const [saturdayTo, setSaturdayTo] = useState(0);
   const [sundayTo, setSundayTo] = useState(0);
-  const history = useHistory();
   const isLoading = useSelector((store) => store.psychologists.isLoading);
   const error = useSelector((store) => store.psychologists.error);
   const query = useQuery();
