--- src/Components/Sessions/TextArea/textArea.module.css
@@ -1,9 +1,0 @@
-.textarea {
-  margin-left: 20px;
-  margin-bottom: 20px;
-  width: 190px;
-  height: 40px;
-  border: 1px solid gray;
-  border-radius: 3px;
-  background: white;
-}
