+++ src/Components/Shared/Modals/ModalDelete/modaldelete.module.css
@@ -1,0 +1,65 @@
+.container {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgb(0,0,0);
+  background-color: rgba(0,0,0,0.4);
+  z-index: 1;
+}
+
+.modal {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  margin: 15% auto;
+  padding-bottom: 1em;
+  width: 50%;
+  height: 40%;
+  border: 0.1em solid #888888;
+  background-color: #FFFFFF;
+}
+
+.header {
+  display: flex;
+  justify-content: space-between;
+  padding: 2em 1em;
+  background-color: #232F3E;
+  color: #FFFFFF;
+  font-size: 20px;
+}
+
+.message {
+  display: flex;
+  justify-content: center;
+  align-self: center;
+  font-size: 30px;
+}
+
+.btnContainer {
+  display: flex;
+  justify-content: center;
+  align-self: center;
+}
+
+.cancelButton {
+  margin-right: 20px;
+  width: 20em;
+  height: 3em;
+  border: none;
+  border-radius: 0.3em;
+  background-color: #232f3e;
+  color: #FFFFFF;
+  cursor: pointer;
+}
+
+.confirmButton {
+  width: 20em;
+  height: 3em;
+  border: none;
+  border-radius: 0.3em;
+  background-color: #FC6225;
+  color: #FFFFFF;
+  cursor: pointer;
+}
